Cambios postgre1

+++ src/main/java/MVC/Entity/Cliente.java
@@ -8,11 +8,11 @@ public class Cliente {
     public Cliente() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -53,7 +53,7 @@ public class Cliente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
     @Column(name = "nombre")
     private String nombre;
     @Column(name = "apellido")
